+++ src/freebsd/utils.rs
@@ -281,16 +281,3 @@ impl std::ops::Deref for KInfoProc {
         &self.0
     }
 }
-
-pub(crate) unsafe fn get_frequency_for_cpu(cpu_nb: c_int) -> u64 {
-    let mut frequency = 0;
-
-    // The information can be missing if it's running inside a VM.
-    if !get_sys_value_by_name(
-        format!("dev.cpu.{cpu_nb}.freq\0").as_bytes(),
-        &mut frequency,
-    ) {
-        frequency = 0;
-    }
-    frequency as _
-}
